Move dependency specs to deps.edn...
...and use lein-tools-deps so that Leiningen will still work.

The main value prop here, as I see it, is that projects using tools.deps
can include eftest via its git repo via SHA, which means that they could
easily use a version of eftest that exists in the repo but has not been
published to Clojars in a release artifact. And that’d apply to _any_
version in the repo — even in a feature branch. It’d also be easy for
projects to use forks of eftest — they’d simply change the git URL in
their deps.edn to point to the fork repo.

For example, #63 was merged to master a week ago, and I’m eager to use
it in my project, but I can’t because a new release with the change
hasn’t yet been published to Clojars. Once _this_ change is merged to
master, I’ll be able to use the very latest code in master without
waiting for it to be published.

The motivation to use lein-tools-deps is to preserve the existing
workflows for running tests, publishing releases, etc, without
duplicating the dependency specifications.